requirements.txt
asgiref==3.4.1
autopep8==1.5.5
Django==3.2.23
django-storages==1.10.1
django-cors-headers == 3.10.1
djangorestframework == 3.14.0
djangorestframework-simplejwt == 4.4.0
jmespath==0.10.0
PyJWT==2.4.0
python-dateutil==2.8.2
six==1.16.0
toml==0.10.2
urllib3==1.26.18
Pillow==8.4.0
pytz==2023.3
sqlparse==0.4.4